+++ module_groups/ecs_fargate_with_elb/variables.tf
@@ -135,3 +135,15 @@ variable "ecs_incoming_allowed_cidr" {
   description = "The CIDR to allow incoming traffic to the ECS tasks.  Overrides the default of the VPC"
   default     = ""
 }
+
+variable "datadog_api_key" {
+  description = "Datadog API key. This will activate the Datadog agent sidecar/replica container. To accommodate additional load, please increase var.cpu to an additional 512 or more. Similarly, increase var.memory to an additional 1024."
+  default     = ""
+  type        = string
+}
+
+variable "datadog_agent_version" {
+  description = "Datadog agent container version."
+  default     = "latest"
+  type        = string
+}